refactor (client): advanced layout
