+++ client/src/routes/(app)/+layout.svelte
@@ -1,0 +1,6 @@
+<script>
+</script>
+
+<div class="flex flex-col h-screen w-96  m-auto border-solid border-2 border-sky-500">
+	<slot />
+</div>
