+++ client/src/routes/(app)/books/+page.svelte
