+++ client/src/routes/(app)/records/+page.svelte
