+++ client/src/routes/(settings)/+layout.svelte
@@ -1,0 +1,8 @@
+<script>
+</script>
+
+<div
+	class="flex flex-col h-screen max-w-md justify-center m-auto border-solid border-2 border-sky-500"
+>
+	<slot />
+</div>
