+++ client/src/routes/+layout.svelte
@@ -2,6 +2,4 @@
 	import '../app.css';
 </script>
 
-<div class="flex flex-col h-screen max-w-md justify-center m-auto border-solid border-2 border-sky-500">
-	<slot />
-</div>
+<slot />
